Add Migration section
This PR proposes to add a Migration section to the RFC template. The purpose of this section is to ensure consideration is given to how an RFC might break existing APIs and workflows so that a discussion can be had in advance of an RFC being approved.

As we get more and more users, this seems like an appropriate step to take to show that stability and continuity are important considerations as we evolve buildpacks.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>